+++ frontend/pages/Homepage/Dashboard/HostsSummary/_styles.scss
@@ -1,4 +1,4 @@
-.host-summary {
+.hosts-summary {
   align-self: center;
   width: 100%;
   max-width: 900px;
